Update token contract bot
