+++ malicious-smart-contract-ml-py/.gitignore
@@ -3,3 +3,6 @@ dist
 forta.config.json
 __pycache__
 .pytest_cache
+malicious_non_token_model_11_05_22.joblib
+malicious_non_token_model_02_07_23_exp2.joblib
+publish.log
