+++ malicious-smart-contract-ml-py/publish.log
@@ -1,10 +1,6 @@
-Wed, 07 Sep 2022 16:17:31 GMT: successfully added agent id 0xeab3b34f9c32e9a5cafb76fccbd98f98f441d9e0499d93c4b476ba754f8f0773 with manifest Qme7UKSvpW89ainQLpfxbRbZcWZNcPFxjkFSuNtYP11cQg
-Wed, 07 Sep 2022 18:58:26 GMT: successfully updated agent id 0xeab3b34f9c32e9a5cafb76fccbd98f98f441d9e0499d93c4b476ba754f8f0773 with manifest QmRFuo7vfuXrDhb8EUoAAb76yAprRBNujoAxDsux8mGQrQ
-Mon, 03 Oct 2022 15:55:26 GMT: successfully pushed image with reference bafybeibsneh2motdsa25fh6kg6bduvuq7yxq6nzbejdabfyilpg7sdo5ye@sha256:d79e442052b8f8683c331a21dc0f7258fc32db9d9a147c571c3a0d392532865b
-Mon, 03 Oct 2022 16:30:03 GMT: successfully pushed image with reference bafybeig2teivto3h7jzxtzhqzpwhxdzdaprwt3sij4izc45eqsiodawxeq@sha256:a93efb4319a37177869254d411d9d9fb20e23f09f66bae91793f5b96e7a2bc23
-Mon, 31 Oct 2022 15:39:33 GMT: successfully pushed image with reference bafybeiezx4dgmvno3a6os3wkuucctwd5vd6q75ox4czvtl3dxyvm445z4e@sha256:8e11f70fd2cccfad4ad210727511c726d0f89ba2de407694b85e2485571b7e9c
-Mon, 31 Oct 2022 16:30:04 GMT: successfully pushed image with reference bafybeiamxlbt447mr25husvmzp4cfj5jxy45dqhkkjttshhucitii3pzxu@sha256:ff9309f0705547a445a660eac4efea4913bc895fea9994b54e7071586888ae4b
-Thu, 22 Dec 2022 19:30:27 GMT: successfully pushed image with reference bafybeib4julgsmiwjfnrykq36secv7q5onue7g5v5h6y6xghjvjmiwkdfq@sha256:d2fd40bf753052104c9d1402f71bdf9e80d974706a90fcbb1fa64dd881e23b13
-Thu, 22 Dec 2022 20:17:49 GMT: successfully pushed image with reference bafybeihlt3jnjpcvtm7k2queatbjlxcmrbgorfrndknyujc2vzqrr5cb4a@sha256:de931736060fd983e0f6679cb821b61e7c2e528731f76786436909d0bf0ffaa2
-Thu, 22 Dec 2022 23:53:00 GMT: successfully pushed image with reference bafybeicztctgcekz6f4ftlua3cc5eg5v4kjngdqsjbd42sunqtsdqrt24i@sha256:d115799d836030b36d86e78bc4a06e42ac58ee1e3f7acbc77bd4b9e3b7592d2f
-Fri, 27 Jan 2023 20:05:26 GMT: successfully pushed image with reference bafybeibf5ejvu76oneyzy76eyuqp3khpojpjv6ib5okz3yd5j6bcl4jpbm@sha256:bba3a5ab7c6c766865780fff5356a46afc480ca0b78e8dffc1f4ff90c59d7f0e
+Thu, 22 Dec 2022 18:28:50 GMT: successfully pushed image with reference bafybeiepws54v5ziiri7rypanepkb62humjc7xl5ywpoqamdyykkcosjxi@sha256:bc4c5e429e27fa7e7e56c4049f5025542636e9c4188d2e059383cc4ba95c8c2b
+Fri, 23 Dec 2022 04:53:39 GMT: successfully pushed image with reference bafybeibisyjragm4rv3xqpw2ln5mzattjtnuywdixccr2cqpx7gdhy5qbi@sha256:e9b7e3abcc68041c48ec7c5af866791fcd97a8b927b07c32fb2621ba779bfdaf
+Fri, 27 Jan 2023 19:50:21 GMT: successfully pushed image with reference bafybeibd6smo2cqeeu46x454rswynyhdqstou4lsccfnnn4gp6qdpbeoh4@sha256:23e72a89a8efa558ccc4fa860a8f842cae41527f4d6b1bb83eb66e77493854d6
+Tue, 07 Feb 2023 19:11:47 GMT: successfully pushed image with reference bafybeihqijl2w4vkfdmsc7a2nczb54enc42flvwh7gpcbpkeozfh47k7re@sha256:990642bbd45d8886d2aa6e9c5fef639967e25bb3c4f7f068c61a1db175627b19
+Thu, 09 Feb 2023 16:45:36 GMT: successfully pushed image with reference bafybeifmyyzhltseesyoxgzkcoy2z5penzglvhcyhu6tagkh4rbqy3bokm@sha256:9473d968d1c49b9117bee0b98452ea539a86d2d8e214d23753b0faadd0145f55
+Thu, 09 Feb 2023 21:49:49 GMT: successfully pushed image with reference bafybeibgxraygs2kxyvhgj5j2n6pybp4yyfcs4lr7suofnwvkc7f7bqjr4@sha256:72b0e6fea05c01cdea2f41149ac81c85268f2b6925b5750f2bd08f98af10d9c8
