--- malicious-smart-contract-ml-v3-py/.dockerignore
@@ -1,3 +1,0 @@
-node_modules
-dist
-forta.config.json
