--- malicious-smart-contract-ml-v3-py/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "malicious-smart-contract-ml-v3",
-  "version": "3.0.0",
-  "description": "Malicious Smart Contract Detection Bot V3",
-  "chainIds": [
-    1, 56, 137, 43114, 42161, 10, 250
-  ],
-  "scripts": {
-    "postinstall": "python3 -m pip install -r requirements_dev.txt",
-    "start": "npm run start:dev",
-    "start:dev": "nodemon --watch src --watch forta.config.json -e py --exec \"forta-agent run\"",
-    "start:prod": "forta-agent run --prod",
-    "tx": "forta-agent run --tx",
-    "block": "forta-agent run --block",
-    "range": "forta-agent run --range",
-    "file": "forta-agent run --file",
-    "publish": "forta-agent publish",
-    "push": "forta-agent push",
-    "disable": "forta-agent disable",
-    "enable": "forta-agent enable",
-    "keyfile": "forta-agent keyfile",
-    "test": "python3 -m pytest"
-  },
-  "dependencies": {
-    "forta-agent": "^0.1.15"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.8"
-  }
-}
