--- malicious-smart-contract-ml-v3-py/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-env =
-    PYTHON_ENV=test
