--- malicious-smart-contract-ml-v3-py/requirements.txt
@@ -1,8 +1,0 @@
-forta_agent>=0.1.9
-setuptools>=61.3.1
-hexbytes>=0.2.2
-evmdasm>=0.1.10
-rlp>=2.0.1
-scikit-learn==1.2.1 # Fixed to avoid bot and model version mismatch.
-
-
