--- malicious-smart-contract-ml-v3-py/requirements_dev.txt
@@ -1,3 +1,0 @@
--r requirements.txt
-pytest==6.2.5
-pytest-env==0.6.2
