--- malicious-smart-contract-ml-v3-py/src/utils.py
@@ -1,108 +1,0 @@
-from hexbytes import HexBytes
-import rlp
-import requests
-from web3 import Web3
-
-
-from src.constants import CONTRACT_SLOT_ANALYSIS_DEPTH, MASK, BOT_ID
-from src.logger import logger
-
-
-def calc_contract_address(w3, address, nonce) -> str:
-    """
-    this function calculates the contract address from sender/nonce
-    :return: contract address: str
-    """
-
-    address_bytes = bytes.fromhex(address[2:].lower())
-    return Web3.toChecksumAddress(Web3.keccak(rlp.encode([address_bytes, nonce]))[-20:])
-
-
-def is_contract(w3, address) -> bool:
-    """
-    this function determines whether address is a contract
-    :return: is_contract: bool
-    """
-    if address is None:
-        return True
-    code = w3.eth.get_code(Web3.toChecksumAddress(address))
-    return code != HexBytes("0x")
-
-
-def get_storage_addresses(w3, address) -> set:
-    """
-    this function returns the addresses that are references in the storage of a contract (first CONTRACT_SLOT_ANALYSIS_DEPTH slots)
-    :return: address_list: list (only returning contract addresses)
-    """
-    if address is None:
-        return set()
-
-    address_set = set()
-    for i in range(CONTRACT_SLOT_ANALYSIS_DEPTH):
-        mem = w3.eth.get_storage_at(Web3.toChecksumAddress(address), i)
-        if mem != HexBytes(
-            "0x0000000000000000000000000000000000000000000000000000000000000000"
-        ):
-            # looking at both areas of the storage slot as - depending on packing - the address could be at the beginning or the end.
-            addr_on_left = mem[0:20].hex()
-            addr_on_right = mem[12:].hex()
-            if is_contract(w3, addr_on_left):
-                address_set.add(Web3.toChecksumAddress(addr_on_left))
-            if is_contract(w3, addr_on_right):
-                address_set.add(Web3.toChecksumAddress(addr_on_right))
-
-    return address_set
-
-
-def get_features(w3, opcodes, contract_creator) -> list:
-    """
-    this function returns the contract opcodes
-    :return: features: list
-    """
-    features = []
-    opcode_addresses = set()
-
-    for i, opcode in enumerate(opcodes):
-        opcode_name = opcode.name
-        # treat unique unknown and invalid opcodes as UNKNOWN OR INVALID
-        if opcode_name.startswith("UNKNOWN") or opcode_name.startswith("INVALID"):
-            opcode_name = opcode.name.split("_")[0]
-        features.append(opcode_name)
-        if len(opcode.operand) == 40 and is_contract(w3, opcode.operand):
-            opcode_addresses.add(Web3.toChecksumAddress(f"0x{opcode.operand}"))
-
-        if opcode_name in {"PUSH4", "PUSH32"}:
-            features.append(opcode.operand)
-        elif opcode_name == "PUSH20":
-            if opcode.operand == contract_creator:
-                features.append("creator")
-            elif opcode.operand == MASK:
-                features.append(MASK)
-            else:
-                features.append("addr")
-
-    features = " ".join(features)
-
-    return features, opcode_addresses
-
-
-def alert_count(chain_id: int, alert_id: str) -> int:
-    alert_stats_url = (
-        f"https://api.forta.network/stats/bot/{BOT_ID}/alerts?chainId={chain_id}"
-    )
-    alert_id_counts = 1
-    alert_counts = 1
-    try:
-        result = requests.get(alert_stats_url).json()
-        alert_id_counts = result["alertIds"][alert_id]["count"]
-        alert_counts = result["total"]["count"]
-    except Exception as err:
-        logger.error(f"Error obtaining alert counts: {err}")
-
-    return alert_id_counts, alert_counts
-
-
-def get_anomaly_score(chain_id: int, alert_id: str) -> float:
-    total_alert_ids, total_alerts = alert_count(chain_id, alert_id)
-
-    return min(total_alert_ids / total_alerts, 1.0)
