--- malicious-smart-contract-ml-v3-py/src/web3_mock.py
@@ -1,90 +1,0 @@
-from hexbytes import HexBytes
-
-EOA_ADDRESS = "0x1c5dCdd006EA78a7E4783f9e6021C32935a10fb4"
-CONTRACT_NO_ADDRESS = "0x04c90C198b2eFF55716079bc06d7CCc4aa4d7512"
-CONTRACT_WITH_ADDRESS = "0x49A9deCA3DcA86aB3A029C2ed629EC8477009Fee"
-CONTRACT = "0x2320A28f52334d62622cc2EaFa15DE55F9987eD9"
-MALICIOUS_CONTRACT = "0x620E59769085F69d2712a53Aa0436821218e1068"
-MALICIOUS_CONTRACT_DEPLOYER = "0x34B21846Ca6bF20264242F01CB8220982084EdBd"
-MALICIOUS_CONTRACT_DEPLOYER_NONCE = 0
-BENIGN_CONTRACT = "0x587969Add789c13F64Bcc34Ff253BD9BFB78f38a"
-SHORT_CONTRACT = "0x587969Add789c13F64Bcc34Ff253BD9BFB78f38b"
-
-
-class Web3Mock:
-    def __init__(self):
-        self.eth = EthMock()
-
-
-class EthMock:
-    def __init__(self):
-        self.contract = ContractMock()
-        self.chain_id = 1
-
-    def get_transaction_count(self, address):
-        if address == EOA_ADDRESS:
-            return 9
-        else:
-            return 0
-
-    def get_storage_at(self, address, position):
-        if address == EOA_ADDRESS:
-            return HexBytes(
-                "0x0000000000000000000000000000000000000000000000000000000000000000"
-            )
-        elif address == CONTRACT_NO_ADDRESS:
-            return HexBytes(
-                "0x0000000000000000000000000000000000000000000000000000000000000005"
-            )
-        elif address == CONTRACT_WITH_ADDRESS:
-            return HexBytes(
-                "0x0000000000000000000000004c90C198b2eFF55716079bc06d7CCc4aa4d7512"
-            )
-        else:
-            return HexBytes(
-                "0x0000000000000000000000000000000000000000000000000000000000000000"
-            )
-
-    def get_code(self, address):
-        if address == CONTRACT_NO_ADDRESS:
-            return HexBytes(
-                "0x608060405234801561001057600080fd5b50610150806100206000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c80632e64cec11461003b5780636057361d14610059575b600080fd5b610043610075565b60405161005091906100d9565b60405180910390f35b610073600480360381019061006e919061009d565b61007e565b005b60008054905090565b8060008190555050565b60008135905061009781610103565b92915050565b6000602082840312156100b3576100b26100fe565b5b60006100c184828501610088565b91505092915050565b6100d3816100f4565b82525050565b60006020820190506100ee60008301846100ca565b92915050565b6000819050919050565b600080fd5b61010c816100f4565b811461011757600080fd5b5056fea26469706673582212209a159a4f3847890f10bfb87871a61eba91c5dbf5ee3cf6398207e292eee22a1664736f6c63430008070033"
-            )
-        elif address == SHORT_CONTRACT:
-            return HexBytes("0x6080604052600436106100ab576")
-        elif address == MALICIOUS_CONTRACT:
-            return HexBytes(
-                "0x6080604052600436106100435760003560e01c80636f307dc31461004c5780639c58769814610077578063c402cd1a14610081578063e184c9be1461008b5761004a565b3661004a57005b005b34801561005857600080fd5b506100616100b6565b60405161006e919061043a565b60405180910390f35b61007f6100d2565b005b610089610269565b005b34801561009757600080fd5b506100a06103ef565b6040516100ad9190610476565b60405180910390f35b60007364aa3364f17a4d01c6f1751fd97c2bd3d7e7f1d5905090565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161461012c57600080fd5b60008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16631e9a695030600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166370a08231727fe7c498a2cf30971ad8f2cbc36bd14ac511566040518263ffffffff1660e01b81526004016101d7919061043a565b602060405180830381865afa1580156101f4573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061021891906104cc565b6040518363ffffffff1660e01b8152600401610235929190610508565b600060405180830381600087803b15801561024f57600080fd5b505af1158015610263573d6000803e3d6000fd5b50505050565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16146102c357600080fd5b600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663a9059cbb33600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166370a08231306040518263ffffffff1660e01b815260040161035d919061043a565b602060405180830381865afa15801561037a573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061039e91906104cc565b6040518363ffffffff1660e01b81526004016103bb929190610508565b600060405180830381600087803b1580156103d557600080fd5b505af11580156103e9573d6000803e3d6000fd5b50505050565b6000610539905090565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000610424826103f9565b9050919050565b61043481610419565b82525050565b600060208201905061044f600083018461042b565b92915050565b600065ffffffffffff82169050919050565b61047081610455565b82525050565b600060208201905061048b6000830184610467565b92915050565b600080fd5b6000819050919050565b6104a981610496565b81146104b457600080fd5b50565b6000815190506104c6816104a0565b92915050565b6000602082840312156104e2576104e1610491565b5b60006104f0848285016104b7565b91505092915050565b61050281610496565b82525050565b600060408201905061051d600083018561042b565b61052a60208301846104f9565b939250505056fea2646970667358221220dc04e9db1101bc3dc7163c383d8cc379c81de83ade1a357cc1ffa73e72590c6864736f6c634300080d0033"
-            )
-        elif address == BENIGN_CONTRACT:
-            return HexBytes(
-                "0x60806040526001600c5534801561001557600080fd5b506040514690806052612d228239604080519182900360520182208282018252600a8352692ab734b9bbb0b8102b1960b11b6020938401528151808301835260018152603160f81b908401528151808401919091527fbfcc8ef98ffbf7b6c3fec7bf5185b566b9863e35a9d83acd49ad6824b5969738818301527fc89efdaa54c0f20c7adf612882df0950f5a951637e0307cdcb4c672f298b8bc6606082015260808101949094523060a0808601919091528151808603909101815260c09094019052825192019190912060035550600580546001600160a01b03191633179055612c1d806101056000396000f3fe608060405234801561001057600080fd5b50600436106101b95760003560e01c80636a627842116100f9578063ba9a7a5611610097578063d21220a711610071578063d21220a7146105da578063d505accf146105e2578063dd62ed3e14610640578063fff6cae91461067b576101b9565b8063ba9a7a5614610597578063bc25cf771461059f578063c45a0155146105d2576101b9565b80637ecebe00116100d35780637ecebe00146104d757806389afcb441461050a57806395d89b4114610556578063a9059cbb1461055e576101b9565b80636a6278421461046957806370a082311461049c5780637464fc3d146104cf576101b9565b806323b872dd116101665780633644e515116101405780633644e51514610416578063485cc9551461041e5780635909c0d5146104595780635a3d549314610461576101b9565b806323b872dd146103ad57806330adf81f146103f0578063313ce567146103f8576101b9565b8063095ea7b311610197578063095ea7b3146103155780630dfe16811461036257806318160ddd14610393576101b9565b8063022c0d9f146101be57806306fdde03146102595780630902f1ac146102d6575b600080fd5b610257600480360360808110156101d457600080fd5b81359160208101359173ffffffffffffffffffffffffffffffffffffffff604083013516919081019060808101606082013564010000000081111561021857600080fd5b82018360208201111561022a57600080fd5b8035906020019184600183028401116401000000008311171561024c57600080fd5b509092509050610683565b005b610261610d57565b6040805160208082528351818301528351919283929083019185019080838360005b8381101561029b578181015183820152602001610283565b50505050905090810190601f1680156102c85780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b6102de610d90565b604080516dffffffffffffffffffffffffffff948516815292909316602083015263ffffffff168183015290519081900360600190f35b61034e6004803603604081101561032b57600080fd5b5073ffffffffffffffffffffffffffffffffffffffff8135169060200135610de5565b604080519115158252519081900360200190f35b61036a610dfc565b6040805173ffffffffffffffffffffffffffffffffffffffff9092168252519081900360200190f35b61039b610e18565b60408051918252519081900360200190f35b61034e600480360360608110156103c357600080fd5b5073ffffffffffffffffffffffffffffffffffffffff813581169160208101359091169060400135610e1e565b61039b610efd565b610400610f21565b6040805160ff9092168252519081900360200190f35b61039b610f26565b6102576004803603604081101561043457600080fd5b5073ffffffffffffffffffffffffffffffffffffffff81358116916020013516610f2c565b61039b611005565b61039b61100b565b61039b6004803603602081101561047f57600080fd5b503573ffffffffffffffffffffffffffffffffffffffff16611011565b61039b600480360360208110156104b257600080fd5b503573ffffffffffffffffffffffffffffffffffffffff166113cb565b61039b6113dd565b61039b600480360360208110156104ed57600080fd5b503573ffffffffffffffffffffffffffffffffffffffff166113e3565b61053d6004803603602081101561052057600080fd5b503573ffffffffffffffffffffffffffffffffffffffff166113f5565b6040805192835260208301919091528051918290030190f35b610261611892565b61034e6004803603604081101561057457600080fd5b5073ffffffffffffffffffffffffffffffffffffffff81351690602001356118cb565b61039b6118d8565b610257600480360360208110156105b557600080fd5b503573ffffffffffffffffffffffffffffffffffffffff166118de565b61036a611ad4565b61036a611af0565b610257600480360360e08110156105f857600080fd5b5073ffffffffffffffffffffffffffffffffffffffff813581169160208101359091169060408101359060608101359060ff6080820135169060a08101359060c00135611b0c565b61039b6004803603604081101561065657600080fd5b5073ffffffffffffffffffffffffffffffffffffffff81358116916020013516611dd8565b610257611df5565b600c546001146106f457604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601160248201527f556e697377617056323a204c4f434b4544000000000000000000000000000000604482015290519081900360640190fd5b6000600c55841515806107075750600084115b61075c576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526025815260200180612b2f6025913960400191505060405180910390fd5b600080610767610d90565b5091509150816dffffffffffffffffffffffffffff168710801561079a5750806dffffffffffffffffffffffffffff1686105b6107ef576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526021815260200180612b786021913960400191505060405180910390fd5b600654600754600091829173ffffffffffffffffffffffffffffffffffffffff91821691908116908916821480159061085457508073ffffffffffffffffffffffffffffffffffffffff168973ffffffffffffffffffffffffffffffffffffffff1614155b6108bf57604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601560248201527f556e697377617056323a20494e56414c49445f544f0000000000000000000000604482015290519081900360640190fd5b8a156108d0576108d0828a8d611fdb565b89156108e1576108e1818a8c611fdb565b86156109c3578873ffffffffffffffffffffffffffffffffffffffff166310d1e85c338d8d8c8c6040518663ffffffff1660e01b8152600401808673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001858152602001848152602001806020018281038252848482818152602001925080828437600081840152601f19601f8201169050808301925050509650505050505050600060405180830381600087803b1580156109aa57600080fd5b505af11580156109be573d6000803e3d6000fd5b505050505b604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905173ffffffffffffffffffffffffffffffffffffffff8416916370a08231916024808301926020929190829003018186803b158015610a2f57600080fd5b505afa158015610a43573d6000803e3d6000fd5b505050506040513d6020811015610a5957600080fd5b5051604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905191955073ffffffffffffffffffffffffffffffffffffffff8316916370a0823191602480820192602092909190829003018186803b158015610acb57600080fd5b505afa158015610adf573d6000803e3d6000fd5b505050506040513d6020811015610af557600080fd5b5051925060009150506dffffffffffffffffffffffffffff85168a90038311610b1f576000610b35565b89856dffffffffffffffffffffffffffff160383035b9050600089856dffffffffffffffffffffffffffff16038311610b59576000610b6f565b89856dffffffffffffffffffffffffffff160383035b90506000821180610b805750600081115b610bd5576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526024815260200180612b546024913960400191505060405180910390fd5b6000610c09610beb84600363ffffffff6121e816565b610bfd876103e863ffffffff6121e816565b9063ffffffff61226e16565b90506000610c21610beb84600363ffffffff6121e816565b9050610c59620f4240610c4d6dffffffffffffffffffffffffffff8b8116908b1663ffffffff6121e816565b9063ffffffff6121e816565b610c69838363ffffffff6121e816565b1015610cd657604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152600c60248201527f556e697377617056323a204b0000000000000000000000000000000000000000604482015290519081900360640190fd5b5050610ce4848488886122e0565b60408051838152602081018390528082018d9052606081018c9052905173ffffffffffffffffffffffffffffffffffffffff8b169133917fd78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d8229181900360800190a350506001600c55505050505050505050565b6040518060400160405280600a81526020017f556e69737761702056320000000000000000000000000000000000000000000081525081565b6008546dffffffffffffffffffffffffffff808216926e0100000000000000000000000000008304909116917c0100000000000000000000000000000000000000000000000000000000900463ffffffff1690565b6000610df233848461259c565b5060015b92915050565b60065473ffffffffffffffffffffffffffffffffffffffff1681565b60005481565b73ffffffffffffffffffffffffffffffffffffffff831660009081526002602090815260408083203384529091528120547fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff14610ee85773ffffffffffffffffffffffffffffffffffffffff84166000908152600260209081526040808320338452909152902054610eb6908363ffffffff61226e16565b73ffffffffffffffffffffffffffffffffffffffff851660009081526002602090815260408083203384529091529020555b610ef384848461260b565b5060019392505050565b7f6e71edae12b1b97f4d1f60370fef10105fa2faae0126114a169c64845d6126c981565b601281565b60035481565b60055473ffffffffffffffffffffffffffffffffffffffff163314610fb257604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601460248201527f556e697377617056323a20464f5242494444454e000000000000000000000000604482015290519081900360640190fd5b6006805473ffffffffffffffffffffffffffffffffffffffff9384167fffffffffffffffffffffffff00000000000000000000000000000000000000009182161790915560078054929093169116179055565b60095481565b600a5481565b6000600c5460011461108457604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601160248201527f556e697377617056323a204c4f434b4544000000000000000000000000000000604482015290519081900360640190fd5b6000600c81905580611094610d90565b50600654604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905193955091935060009273ffffffffffffffffffffffffffffffffffffffff909116916370a08231916024808301926020929190829003018186803b15801561110e57600080fd5b505afa158015611122573d6000803e3d6000fd5b505050506040513d602081101561113857600080fd5b5051600754604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905192935060009273ffffffffffffffffffffffffffffffffffffffff909216916370a0823191602480820192602092909190829003018186803b1580156111b157600080fd5b505afa1580156111c5573d6000803e3d6000fd5b505050506040513d60208110156111db57600080fd5b505190506000611201836dffffffffffffffffffffffffffff871663ffffffff61226e16565b90506000611225836dffffffffffffffffffffffffffff871663ffffffff61226e16565b9050600061123387876126ec565b600054909150806112705761125c6103e8610bfd611257878763ffffffff6121e816565b612878565b985061126b60006103e86128ca565b6112cd565b6112ca6dffffffffffffffffffffffffffff8916611294868463ffffffff6121e816565b8161129b57fe5b046dffffffffffffffffffffffffffff89166112bd868563ffffffff6121e816565b816112c457fe5b0461297a565b98505b60008911611326576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526028815260200180612bc16028913960400191505060405180910390fd5b6113308a8a6128ca565b61133c86868a8a6122e0565b811561137e5760085461137a906dffffffffffffffffffffffffffff808216916e01000000000000000000000000000090041663ffffffff6121e816565b600b555b6040805185815260208101859052815133927f4c209b5fc8ad50758f13e2e1088ba56a560dff690a1c6fef26394f4c03821c4f928290030190a250506001600c5550949695505050505050565b60016020526000908152604090205481565b600b5481565b60046020526000908152604090205481565b600080600c5460011461146957604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601160248201527f556e697377617056323a204c4f434b4544000000000000000000000000000000604482015290519081900360640190fd5b6000600c81905580611479610d90565b50600654600754604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905194965092945073ffffffffffffffffffffffffffffffffffffffff9182169391169160009184916370a08231916024808301926020929190829003018186803b1580156114fb57600080fd5b505afa15801561150f573d6000803e3d6000fd5b505050506040513d602081101561152557600080fd5b5051604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905191925060009173ffffffffffffffffffffffffffffffffffffffff8516916370a08231916024808301926020929190829003018186803b15801561159957600080fd5b505afa1580156115ad573d6000803e3d6000fd5b505050506040513d60208110156115c357600080fd5b5051306000908152600160205260408120549192506115e288886126ec565b600054909150806115f9848763ffffffff6121e816565b8161160057fe5b049a5080611614848663ffffffff6121e816565b8161161b57fe5b04995060008b11801561162e575060008a115b611683576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526028815260200180612b996028913960400191505060405180910390fd5b61168d3084612992565b611698878d8d611fdb565b6116a3868d8c611fdb565b604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905173ffffffffffffffffffffffffffffffffffffffff8916916370a08231916024808301926020929190829003018186803b15801561170f57600080fd5b505afa158015611723573d6000803e3d6000fd5b505050506040513d602081101561173957600080fd5b5051604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905191965073ffffffffffffffffffffffffffffffffffffffff8816916370a0823191602480820192602092909190829003018186803b1580156117ab57600080fd5b505afa1580156117bf573d6000803e3d6000fd5b505050506040513d60208110156117d557600080fd5b505193506117e585858b8b6122e0565b811561182757600854611823906dffffffffffffffffffffffffffff808216916e01000000000000000000000000000090041663ffffffff6121e816565b600b555b604080518c8152602081018c9052815173ffffffffffffffffffffffffffffffffffffffff8f169233927fdccd412f0b1252819cb1fd330b93224ca42612892bb3f4f789976e6d81936496929081900390910190a35050505050505050506001600c81905550915091565b6040518060400160405280600681526020017f554e492d5632000000000000000000000000000000000000000000000000000081525081565b6000610df233848461260b565b6103e881565b600c5460011461194f57604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601160248201527f556e697377617056323a204c4f434b4544000000000000000000000000000000604482015290519081900360640190fd5b6000600c55600654600754600854604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905173ffffffffffffffffffffffffffffffffffffffff9485169490931692611a2b9285928792611a26926dffffffffffffffffffffffffffff169185916370a0823191602480820192602092909190829003018186803b1580156119ee57600080fd5b505afa158015611a02573d6000803e3d6000fd5b505050506040513d6020811015611a1857600080fd5b50519063ffffffff61226e16565b611fdb565b600854604080517f70a082310000000000000000000000000000000000000000000000000000000081523060048201529051611aca9284928792611a26926e01000000000000000000000000000090046dffffffffffffffffffffffffffff169173ffffffffffffffffffffffffffffffffffffffff8616916370a0823191602480820192602092909190829003018186803b1580156119ee57600080fd5b50506001600c5550565b60055473ffffffffffffffffffffffffffffffffffffffff1681565b60075473ffffffffffffffffffffffffffffffffffffffff1681565b42841015611b7b57604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601260248201527f556e697377617056323a20455850495245440000000000000000000000000000604482015290519081900360640190fd5b60035473ffffffffffffffffffffffffffffffffffffffff80891660008181526004602090815260408083208054600180820190925582517f6e71edae12b1b97f4d1f60370fef10105fa2faae0126114a169c64845d6126c98186015280840196909652958d166060860152608085018c905260a085019590955260c08085018b90528151808603909101815260e0850182528051908301207f19010000000000000000000000000000000000000000000000000000000000006101008601526101028501969096526101228085019690965280518085039096018652610142840180825286519683019690962095839052610162840180825286905260ff89166101828501526101a284018890526101c28401879052519193926101e2808201937fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe081019281900390910190855afa158015611cdc573d6000803e3d6000fd5b50506040517fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe0015191505073ffffffffffffffffffffffffffffffffffffffff811615801590611d5757508873ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff16145b611dc257604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601c60248201527f556e697377617056323a20494e56414c49445f5349474e415455524500000000604482015290519081900360640190fd5b611dcd89898961259c565b505050505050505050565b600260209081526000928352604080842090915290825290205481565b600c54600114611e6657604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601160248201527f556e697377617056323a204c4f434b4544000000000000000000000000000000604482015290519081900360640190fd5b6000600c55600654604080517f70a082310000000000000000000000000000000000000000000000000000000081523060048201529051611fd49273ffffffffffffffffffffffffffffffffffffffff16916370a08231916024808301926020929190829003018186803b158015611edd57600080fd5b505afa158015611ef1573d6000803e3d6000fd5b505050506040513d6020811015611f0757600080fd5b5051600754604080517f70a08231000000000000000000000000000000000000000000000000000000008152306004820152905173ffffffffffffffffffffffffffffffffffffffff909216916370a0823191602480820192602092909190829003018186803b158015611f7a57600080fd5b505afa158015611f8e573d6000803e3d6000fd5b505050506040513d6020811015611fa457600080fd5b50516008546dffffffffffffffffffffffffffff808216916e0100000000000000000000000000009004166122e0565b6001600c55565b604080518082018252601981527f7472616e7366657228616464726573732c75696e743235362900000000000000602091820152815173ffffffffffffffffffffffffffffffffffffffff85811660248301526044808301869052845180840390910181526064909201845291810180517bffffffffffffffffffffffffffffffffffffffffffffffffffffffff167fa9059cbb000000000000000000000000000000000000000000000000000000001781529251815160009460609489169392918291908083835b602083106120e157805182527fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe090920191602091820191016120a4565b6001836020036101000a0380198251168184511680821785525050505050509050019150506000604051808303816000865af19150503d8060008114612143576040519150601f19603f3d011682016040523d82523d6000602084013e612148565b606091505b5091509150818015612176575080511580612176575080806020019051602081101561217357600080fd5b50515b6121e157604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601a60248201527f556e697377617056323a205452414e534645525f4641494c4544000000000000604482015290519081900360640190fd5b5050505050565b60008115806122035750508082028282828161220057fe5b04145b610df657604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601460248201527f64732d6d6174682d6d756c2d6f766572666c6f77000000000000000000000000604482015290519081900360640190fd5b80820382811115610df657604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601560248201527f64732d6d6174682d7375622d756e646572666c6f770000000000000000000000604482015290519081900360640190fd5b6dffffffffffffffffffffffffffff841180159061230c57506dffffffffffffffffffffffffffff8311155b61237757604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601360248201527f556e697377617056323a204f564552464c4f5700000000000000000000000000604482015290519081900360640190fd5b60085463ffffffff428116917c0100000000000000000000000000000000000000000000000000000000900481168203908116158015906123c757506dffffffffffffffffffffffffffff841615155b80156123e257506dffffffffffffffffffffffffffff831615155b15612492578063ffffffff16612425856123fb86612a57565b7bffffffffffffffffffffffffffffffffffffffffffffffffffffffff169063ffffffff612a7b16565b600980547bffffffffffffffffffffffffffffffffffffffffffffffffffffffff929092169290920201905563ffffffff8116612465846123fb87612a57565b600a80547bffffffffffffffffffffffffffffffffffffffffffffffffffffffff92909216929092020190555b600880547fffffffffffffffffffffffffffffffffffff0000000000000000000000000000166dffffffffffffffffffffffffffff888116919091177fffffffff0000000000000000000000000000ffffffffffffffffffffffffffff166e0100000000000000000000000000008883168102919091177bffffffffffffffffffffffffffffffffffffffffffffffffffffffff167c010000000000000000000000000000000000000000000000000000000063ffffffff871602179283905560408051848416815291909304909116602082015281517f1c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad1929181900390910190a1505050505050565b73ffffffffffffffffffffffffffffffffffffffff808416600081815260026020908152604080832094871680845294825291829020859055815185815291517f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b9259281900390910190a3505050565b73ffffffffffffffffffffffffffffffffffffffff8316600090815260016020526040902054612641908263ffffffff61226e16565b73ffffffffffffffffffffffffffffffffffffffff8085166000908152600160205260408082209390935590841681522054612683908263ffffffff612abc16565b73ffffffffffffffffffffffffffffffffffffffff80841660008181526001602090815260409182902094909455805185815290519193928716927fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef92918290030190a3505050565b600080600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663017e7e586040518163ffffffff1660e01b815260040160206040518083038186803b15801561275757600080fd5b505afa15801561276b573d6000803e3d6000fd5b505050506040513d602081101561278157600080fd5b5051600b5473ffffffffffffffffffffffffffffffffffffffff821615801594509192509061286457801561285f5760006127d86112576dffffffffffffffffffffffffffff88811690881663ffffffff6121e816565b905060006127e583612878565b90508082111561285c576000612813612804848463ffffffff61226e16565b6000549063ffffffff6121e816565b905060006128388361282c86600563ffffffff6121e816565b9063ffffffff612abc16565b9050600081838161284557fe5b04905080156128585761285887826128ca565b5050505b50505b612870565b8015612870576000600b555b505092915050565b600060038211156128bb575080600160028204015b818110156128b5578091506002818285816128a457fe5b0401816128ad57fe5b04905061288d565b506128c5565b81156128c5575060015b919050565b6000546128dd908263ffffffff612abc16565b600090815573ffffffffffffffffffffffffffffffffffffffff8316815260016020526040902054612915908263ffffffff612abc16565b73ffffffffffffffffffffffffffffffffffffffff831660008181526001602090815260408083209490945583518581529351929391927fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef9281900390910190a35050565b6000818310612989578161298b565b825b9392505050565b73ffffffffffffffffffffffffffffffffffffffff82166000908152600160205260409020546129c8908263ffffffff61226e16565b73ffffffffffffffffffffffffffffffffffffffff831660009081526001602052604081209190915554612a02908263ffffffff61226e16565b600090815560408051838152905173ffffffffffffffffffffffffffffffffffffffff8516917fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef919081900360200190a35050565b6dffffffffffffffffffffffffffff166e0100000000000000000000000000000290565b60006dffffffffffffffffffffffffffff82167bffffffffffffffffffffffffffffffffffffffffffffffffffffffff841681612ab457fe5b049392505050565b80820182811015610df657604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601460248201527f64732d6d6174682d6164642d6f766572666c6f77000000000000000000000000604482015290519081900360640190fdfe556e697377617056323a20494e53554646494349454e545f4f55545055545f414d4f554e54556e697377617056323a20494e53554646494349454e545f494e5055545f414d4f554e54556e697377617056323a20494e53554646494349454e545f4c4951554944495459556e697377617056323a20494e53554646494349454e545f4c49515549444954595f4255524e4544556e697377617056323a20494e53554646494349454e545f4c49515549444954595f4d494e544544a265627a7a723158207dca18479e58487606bf70c79e44d8dee62353c9ee6d01f9a9d70885b8765f2264736f6c63430005100032454950373132446f6d61696e28737472696e67206e616d652c737472696e672076657273696f6e2c75696e7432353620636861696e49642c6164647265737320766572696679696e67436f6e747261637429"
-            )
-        elif address == CONTRACT:
-            return HexBytes(
-                "0x608060405234801561001057600080fd5b50610150806100206000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c80632e64cec11461003b5780636057361d14610059575b600080fd5b610043610075565b60405161005091906100d9565b60405180910390f35b610073600480360381019061006e919061009d565b61007e565b005b60008054905090565b8060008190555050565b60008135905061009781610103565b92915050565b6000602082840312156100b3576100b26100fe565b5b60006100c184828501610088565b91505092915050565b6100d3816100f4565b82525050565b60006020820190506100ee60008301846100ca565b92915050565b6000819050919050565b600080fd5b61010c816100f4565b811461011757600080fd5b5056fea26469706673582212209a159a4f3847890f10bfb87871a61eba91c5dbf5ee3cf6398207e292eee22a1664736f6c63430008070033"
-            )
-        elif address == CONTRACT_WITH_ADDRESS:
-            return HexBytes(
-                "0x6080604052600436106100595760003560e01c806310d1e85c146100655780632b6a3f7e1461008e57806375ab9782146100be578063dd869c35146100e7578063f0136a6314610110578063f23a6e611461012757610060565b3661006057005b600080fd5b34801561007157600080fd5b5061008c60048036038101906100879190611659565b610164565b005b6100a860048036038101906100a39190611486565b6108b9565b6040516100b59190611b95565b60405180910390f35b3480156100ca57600080fd5b506100e560048036038101906100e091906114fa565b610a64565b005b3480156100f357600080fd5b5061010e6004803603810190610109919061173b565b610a6c565b005b34801561011c57600080fd5b5061012561106f565b005b34801561013357600080fd5b5061014e600480360381019061014991906115bf565b611177565b60405161015b9190611b7a565b60405180910390f35b6000604051806101200160405280600f60065460048110610188576101876120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001600073ffffffffffffffffffffffffffffffffffffffff168152602001600081526020016000815260200160008152602001600081526020016000151581526020016001151581526020016000151581525090506000604051806101200160405280600f6006546004811061023d5761023c6120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001600073ffffffffffffffffffffffffffffffffffffffff168152602001600081526020016000815260200160008152602001600081526020016000151581526020016001151581526020016000151581525090506000600167ffffffffffffffff8111156102eb576102ea6120dd565b5b6040519080825280602002602001820160405280156103195781602001602082028036833780820191505090505b5090506000600167ffffffffffffffff811115610339576103386120dd565b5b6040519080825280602002602001820160405280156103675781602001602082028036833780820191505090505b5090506000600167ffffffffffffffff811115610387576103866120dd565b5b6040519080825280602002602001820160405280156103b55781602001602082028036833780820191505090505b50905030836000815181106103cd576103cc6120ae565b5b602002602001019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff168152505060028260008151811061041d5761041c6120ae565b5b602002602001018181525050600554816000815181106104405761043f6120ae565b5b6020026020010181815250506000601860009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166302e236bc3086868a6040518563ffffffff1660e01b81526004016104af9493929190611aea565b602060405180830381600087803b1580156104c957600080fd5b505af11580156104dd573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190610501919061170e565b90506000601860009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166302e236bc3087868a6040518563ffffffff1660e01b81526004016105669493929190611aea565b602060405180830381600087803b15801561058057600080fd5b505af1158015610594573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906105b8919061170e565b9050601860009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663f2465f09828560008151811061060d5761060c6120ae565b5b602002602001015160016106219190611dcb565b6040518363ffffffff1660e01b815260040161063e929190611c6b565b600060405180830381600087803b15801561065857600080fd5b505af115801561066c573d6000803e3d6000fd5b5050505060008b141561079457600f6006546004811061068f5761068e6120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663a9059cbb6014600654600481106106e2576106e16120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff166103e861044c8e6107159190611e52565b61071f9190611e21565b6040518363ffffffff1660e01b815260040161073c929190611b51565b602060405180830381600087803b15801561075657600080fd5b505af115801561076a573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061078e91906116e1565b506108ab565b600f600654600481106107aa576107a96120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663a9059cbb6014600654600481106107fd576107fc6120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff166103e861044c8f6108309190611e52565b61083a9190611e21565b6040518363ffffffff1660e01b8152600401610857929190611b51565b602060405180830381600087803b15801561087157600080fd5b505af1158015610885573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906108a991906116e1565b505b505050505050505050505050565b606060008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614806109625750600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16145b6109a1576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161099890611c01565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168573ffffffffffffffffffffffffffffffffffffffff1614156109db57600080fd5b6000808673ffffffffffffffffffffffffffffffffffffffff16868686604051610a06929190611a7f565b60006040518083038185875af1925050503d8060008114610a43576040519150601f19603f3d011682016040523d82523d6000602084013e610a48565b606091505b509150915081610a5757600080fd5b8092505050949350505050565b505050505050565b60008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161480610b135750600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16145b610b52576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610b4990611c01565b60405180910390fd5b816004819055508060058190555082600681905550731820a4b7618bde71dce8cdc73aab6c95905fad2473ffffffffffffffffffffffffffffffffffffffff166329965a1d307f29ddb589b1fb5fc7cf394961c1adf5f8c6454761adf795e67fe149f658abe895306040518463ffffffff1660e01b8152600401610bd893929190611ab3565b600060405180830381600087803b158015610bf257600080fd5b505af1158015610c06573d6000803e3d6000fd5b50505050600f60065460048110610c2057610c1f6120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663095ea7b3732320a28f52334d62622cc2eafa15de55f9987ed97fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff6040518363ffffffff1660e01b8152600401610cb0929190611b51565b602060405180830381600087803b158015610cca57600080fd5b505af1158015610cde573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190610d0291906116e1565b50601360065460048110610d1957610d186120ae565b5b602091828204019190069054906101000a900460ff1615610dee57601460065460048110610d4a57610d496120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663022c0d9f6004546002610d969190611e52565b6000306040518463ffffffff1660e01b8152600401610db793929190611c21565b600060405180830381600087803b158015610dd157600080fd5b505af1158015610de5573d6000803e3d6000fd5b50505050610ea4565b601460065460048110610e0457610e036120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663022c0d9f60006004546002610e529190611e52565b306040518463ffffffff1660e01b8152600401610e7193929190611bb7565b600060405180830381600087803b158015610e8b57600080fd5b505af1158015610e9f573d6000803e3d6000fd5b505050505b600f60065460048110610eba57610eb96120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663a9059cbb600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16600f60065460048110610f3057610f2f6120ae565b5b0160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166370a08231306040518263ffffffff1660e01b8152600401610f8a9190611a98565b60206040518083038186803b158015610fa257600080fd5b505afa158015610fb6573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190610fda919061170e565b6040518363ffffffff1660e01b8152600401610ff7929190611b51565b602060405180830381600087803b15801561101157600080fd5b505af1158015611025573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061104991906116e1565b506000600381905550600060048190555060006005819055506000600681905550505050565b60008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614806111165750600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16145b611155576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161114c90611c01565b60405180910390fd5b6000600381905550600060048190555060006005819055506000600681905550565b60008060035414156111a0576003600081548092919061119690612007565b919050555061132a565b6001600354141561132857600360008154809291906111be90612007565b9190505550601860009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16635dcb7ab26001601960009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663bc9683266040518163ffffffff1660e01b815260040160206040518083038186803b15801561126b57600080fd5b505afa15801561127f573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906112a3919061170e565b6112ad9190611eac565b60045460016040518463ffffffff1660e01b81526004016112d093929190611c94565b602060405180830381600087803b1580156112ea57600080fd5b505af11580156112fe573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190611322919061170e565b50611329565b5b5b7ff23a6e612e1ff4830e658fe43f4e3cb4a5f8170bd5d9e69fb5d7a7fa9e4fdf9790509695505050505050565b600061136a61136584611cf0565b611ccb565b9050828152602081018484840111156113865761138561211b565b5b611391848285611f94565b509392505050565b6000813590506113a881612190565b92915050565b6000813590506113bd816121a7565b92915050565b6000815190506113d2816121be565b92915050565b60008083601f8401126113ee576113ed612111565b5b8235905067ffffffffffffffff81111561140b5761140a61210c565b5b60208301915083600182028301111561142757611426612116565b5b9250929050565b600082601f83011261144357611442612111565b5b8135611453848260208601611357565b91505092915050565b60008135905061146b816121d5565b92915050565b600081519050611480816121d5565b92915050565b600080600080606085870312156114a05761149f612125565b5b60006114ae878288016113ae565b94505060206114bf8782880161145c565b935050604085013567ffffffffffffffff8111156114e0576114df612120565b5b6114ec878288016113d8565b925092505092959194509250565b60008060008060008060c0878903121561151757611516612125565b5b600061152589828a01611399565b965050602061153689828a01611399565b955050604061154789828a01611399565b945050606061155889828a0161145c565b935050608087013567ffffffffffffffff81111561157957611578612120565b5b61158589828a0161142e565b92505060a087013567ffffffffffffffff8111156115a6576115a5612120565b5b6115b289828a0161142e565b9150509295509295509295565b60008060008060008060a087890312156115dc576115db612125565b5b60006115ea89828a01611399565b96505060206115fb89828a01611399565b955050604061160c89828a0161145c565b945050606061161d89828a0161145c565b935050608087013567ffffffffffffffff81111561163e5761163d612120565b5b61164a89828a016113d8565b92509250509295509295509295565b60008060008060006080868803121561167557611674612125565b5b600061168388828901611399565b95505060206116948882890161145c565b94505060406116a58882890161145c565b935050606086013567ffffffffffffffff8111156116c6576116c5612120565b5b6116d2888289016113d8565b92509250509295509295909350565b6000602082840312156116f7576116f6612125565b5b6000611705848285016113c3565b91505092915050565b60006020828403121561172457611723612125565b5b600061173284828501611471565b91505092915050565b60008060006060848603121561175457611753612125565b5b60006117628682870161145c565b93505060206117738682870161145c565b92505060406117848682870161145c565b9150509250925092565b600061179a83836117be565b60208301905092915050565b60006117b28383611a61565b60208301905092915050565b6117c781611ee0565b82525050565b6117d681611ee0565b82525050565b60006117e782611d41565b6117f18185611d7c565b93506117fc83611d21565b8060005b8381101561182d578151611814888261178e565b975061181f83611d62565b925050600181019050611800565b5085935050505092915050565b600061184582611d4c565b61184f8185611d8d565b935061185a83611d31565b8060005b8381101561188b57815161187288826117a6565b975061187d83611d6f565b92505060018101905061185e565b5085935050505092915050565b6118a181611f04565b82525050565b6118b081611f10565b82525050565b6118bf81611f1a565b82525050565b60006118d18385611daf565b93506118de838584611f94565b82840190509392505050565b60006118f582611d57565b6118ff8185611d9e565b935061190f818560208601611fa3565b6119188161212a565b840191505092915050565b61192c81611f70565b82525050565b61193b81611f82565b82525050565b600061194e600183611d9e565b91506119598261213b565b602082019050919050565b6000611971600083611d9e565b915061197c82612164565b600082019050919050565b6000611994600e83611dba565b915061199f82612167565b602082019050919050565b610120820160008201516119c160008501826117be565b5060208201516119d460208501826117be565b5060408201516119e76040850182611a61565b5060608201516119fa6060850182611a61565b506080820151611a0d6080850182611a61565b5060a0820151611a2060a0850182611a61565b5060c0820151611a3360c0850182611898565b5060e0820151611a4660e0850182611898565b50610100820151611a5b610100850182611898565b50505050565b611a6a81611f66565b82525050565b611a7981611f66565b82525050565b6000611a8c8284866118c5565b91508190509392505050565b6000602082019050611aad60008301846117cd565b92915050565b6000606082019050611ac860008301866117cd565b611ad560208301856118a7565b611ae260408301846117cd565b949350505050565b60006101a082019050611b0060008301876117cd565b8181036020830152611b1181611964565b90508181036040830152611b2581866117dc565b90508181036060830152611b39818561183a565b9050611b4860808301846119aa565b95945050505050565b6000604082019050611b6660008301856117cd565b611b736020830184611a70565b9392505050565b6000602082019050611b8f60008301846118b6565b92915050565b60006020820190508181036000830152611baf81846118ea565b905092915050565b6000608082019050611bcc6000830186611923565b611bd96020830185611a70565b611be660408301846117cd565b8181036060830152611bf781611941565b9050949350505050565b60006020820190508181036000830152611c1a81611987565b9050919050565b6000608082019050611c366000830186611a70565b611c436020830185611923565b611c5060408301846117cd565b8181036060830152611c6181611941565b9050949350505050565b6000604082019050611c806000830185611a70565b611c8d6020830184611a70565b9392505050565b6000606082019050611ca96000830186611a70565b611cb66020830185611a70565b611cc36040830184611932565b949350505050565b6000611cd5611ce6565b9050611ce18282611fd6565b919050565b6000604051905090565b600067ffffffffffffffff821115611d0b57611d0a6120dd565b5b611d148261212a565b9050602081019050919050565b6000819050602082019050919050565b6000819050602082019050919050565b600081519050919050565b600081519050919050565b600081519050919050565b6000602082019050919050565b6000602082019050919050565b600082825260208201905092915050565b600082825260208201905092915050565b600082825260208201905092915050565b600081905092915050565b600082825260208201905092915050565b6000611dd682611f66565b9150611de183611f66565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff03821115611e1657611e15612050565b5b828201905092915050565b6000611e2c82611f66565b9150611e3783611f66565b925082611e4757611e4661207f565b5b828204905092915050565b6000611e5d82611f66565b9150611e6883611f66565b9250817fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0483118215151615611ea157611ea0612050565b5b828202905092915050565b6000611eb782611f66565b9150611ec283611f66565b925082821015611ed557611ed4612050565b5b828203905092915050565b6000611eeb82611f46565b9050919050565b6000611efd82611f46565b9050919050565b60008115159050919050565b6000819050919050565b60007fffffffff0000000000000000000000000000000000000000000000000000000082169050919050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000819050919050565b6000611f7b82611f66565b9050919050565b6000611f8d82611f66565b9050919050565b82818337600083830152505050565b60005b83811015611fc1578082015181840152602081019050611fa6565b83811115611fd0576000848401525b50505050565b611fdf8261212a565b810181811067ffffffffffffffff82111715611ffe57611ffd6120dd565b5b80604052505050565b600061201282611f66565b91507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff82141561204557612044612050565b5b600182019050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052603260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b600080fd5b600080fd5b600080fd5b600080fd5b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f7800000000000000000000000000000000000000000000000000000000000000600082015250565b50565b7f4e6f7420617574686f72697a6564000000000000000000000000000000000000600082015250565b61219981611ee0565b81146121a457600080fd5b50565b6121b081611ef2565b81146121bb57600080fd5b50565b6121c781611f04565b81146121d257600080fd5b50565b6121de81611f66565b81146121e957600080fd5b5056fea26469706673582212208b55d100ed0d358508a8d8a2ce6cd55b751fc629c49e2d8ee10e62825ac0a18e64736f6c63430008060033"
-            )
-        elif address == EOA_ADDRESS:
-            return HexBytes("0x")
-        else:
-            return HexBytes("0x")
-
-
-class ContractMock:
-    def __init__(self):
-        self.functions = FunctionsMock()
-
-    def __call__(self, address, *args, **kwargs):
-        return self
-
-
-class FunctionsMock:
-    def __init__(self):
-        self.return_value = None
-
-    def call(self, *_, **__):
-        return self.return_value
