+++ malicious-token-contract-ml-py/.gitignore
@@ -4,4 +4,5 @@ forta.config.json
 __pycache__
 .pytest_cache
 malicious_token_model_sighashes_10_29_22.joblib
+malicious_token_model_02_07_23_exp6.joblib
