+++ malicious-token-contract-ml-py/Dockerfile
@@ -19,7 +19,7 @@ ENV NODE_ENV=production
 # Uncomment the following line to enable agent logging
 LABEL "network.forta.settings.agent-logs.enable"="true"
 WORKDIR /app
-COPY ./malicious_token_model_sighashes_10_29_22.joblib ./
+COPY ./malicious_token_model_02_07_23_exp6.joblib ./
 COPY ./src ./src
 COPY package*.json ./
 RUN npm ci --production
