+++ malicious-token-contract-ml-py/README.md
@@ -7,7 +7,6 @@ This detection bot detects when a suspicious (erc20/erc721/erc1155/erc777) token
 ### Model Configuration
 
 *Data Used For Training*
-- Created date: 10/29/2022
 - Trained on 88,368 benign and 391 malicious token contracts. Datasets can be found at [forta-network/labelled-datasets](https://github.com/forta-network/labelled-datasets)
 
 *Algorithm*
@@ -22,9 +21,20 @@ This technique is called TF-IDF (term frequency–inverse document frequency), a
   * Analyzing in chunks helps retain the relative position information of the smart contract opcodes.
 * ML Model: SGD Classifier with loss set to “log_loss”
 
-*Performance*
+**Model Versions**
 
-Using stratified 5-fold cross validation and decision threshold=0.5, the model predicted malicious contracts with average precision=84.1% and recall=38.1%.
+| **Model Version** | V1         | V2                        |                      |   |   |
+|-------------------|------------|---------------------------|----------------------|---|---|
+|  **Created Date** | 10/29/2022 |      02/07/2023           |                      |   |   |
+| **Avg Precision** | 84.1%      | 78.4753%                  |                      |   |   |
+| **Avg Recall**    | 38.1%      | 52.358%                   |                      |   |   |
+| **Avg F1-Score**  | 49.57%     | 61.241%                   |                      |   |   |
+| **Alert Rate**    | 19         | TODO                      |                      |   |   |
+| **Notes**         |            |                           |                      |   |   |
+
+
+* Average precision and recall were calculated via stratified 5-fold cross validation with decision threshold set to `0.5`
+* Alert-rate = number of ethereum alerts daily (avg of 7 days)
 
 ## Supported Chains
 
@@ -51,7 +61,6 @@ The agent behaviour can be verified with the following transactions:
 
 Ethereum Mainnet Transactions
 
-- 0xcc915c5eece83e7c3d2d696a8893233b9da880716cb43d145dcab5c80344c741 (fake USDT erc20 Token - Fake_Phishing5959)
 - 0x664b60a3b78d49a5d2787ef4ea7a599660777003404857da3b3e6c9f103c2ad1 (HydrogenBlue (HydroB) erc20 Token Phishing Scam - Fake_Phishing2600)
 - 0x9b8e791189d4fb3a4b30066c8a8cdd625ef8830d942713ddf8c6221677220cf1 (Token impersonation of INVSBLE erc721 token by Invisible Friends- Fake_Phishing5349)
 - 0x68d2f2a7fbdd6775c48ad010c7842f840012e9ac838b9d894e8188ae5c5401b1 (Bunny Buddies erc721 Token Phishing Scam - Fake_Phishing5328)
