+++ malicious-token-contract-ml-py/src/findings.py
@@ -1,34 +1,58 @@
 from forta_agent import Finding, FindingType, FindingSeverity
 
 
-class MaliciousTokenContractFindings:
-    @staticmethod
-    def malicious_contract_creation(
+class TokenContractFindings:
+    def __init__(
+        self,
         from_address: str,
-        token_type: str,
         contract_address: str,
         contained_addresses: set,
         model_score: float,
         model_threshold: float,
-        anomaly_score: float,
-        labels: list
-    ) -> Finding:
-        metadata = {
+    ):
+        self.metadata = {
             "address_contained_in_created_contract_" + str(i): address
             for i, address in enumerate(contained_addresses, 1)
         }
-        metadata["model_score"] = str(model_score)
-        metadata["model_threshold"] = str(model_threshold)
-        metadata["anomaly_score"] = anomaly_score
+        self.metadata["model_score"] = str(model_score)
+        self.metadata["model_threshold"] = str(model_threshold)
+        self.description = f"{from_address} created contract {contract_address}"
+        self.labels = []
 
+    def malicious_contract_creation(
+        self,
+        anomaly_score: float,
+        labels: list,
+    ) -> Finding:
+        self.metadata["anomaly_score"] = anomaly_score
+        self.label = labels
         return Finding(
             {
                 "name": "Suspicious Token Contract Creation",
-                "description": f"{from_address} created {token_type}-like contract {contract_address}",
+                "description": self.description,
                 "alert_id": "SUSPICIOUS-TOKEN-CONTRACT-CREATION",
                 "type": FindingType.Suspicious,
                 "severity": FindingSeverity.High,
-                "metadata": metadata,
-                "labels": labels
+                "metadata": self.metadata,
+                "labels": self.labels,
+            }
+        )
+
+    def safe_contract_creation(
+        self,
+        anomaly_score: float,
+        labels: list,
+    ) -> Finding:
+        self.metadata["anomaly_score"] = anomaly_score
+        self.label = labels
+        return Finding(
+            {
+                "name": "Safe Token Contract Creation",
+                "description": self.description,
+                "alert_id": "SAFE-TOKEN-CONTRACT-CREATION",
+                "type": FindingType.Info,
+                "severity": FindingSeverity.Info,
+                "metadata": self.metadata,
+                "labels": self.labels,
             }
         )
