+++ malicious-token-contract-ml-py/src/web3_mock.py
@@ -54,7 +54,7 @@ class EthMock:
             return HexBytes("0x6080604052600436106100ab576")
         elif address == MALICIOUS_TOKEN_CONTRACT:
             return HexBytes(
-                "0x60c060405260066080819052652a32ba3432b960d11b60a09081526100279160009190610097565b50604080518082019091526004808252631554d11560e21b602090920191825261005391600191610097565b506c0b5c0e8d21d902d61fa00000006002556003805460ff1916601217905534801561007e57600080fd5b506002543360009081526004602052604090205561016b565b8280546100a390610130565b90600052602060002090601f0160209004810192826100c5576000855561010b565b82601f106100de57805160ff191683800117855561010b565b8280016001018555821561010b579182015b8281111561010b5782518255916020019190600101906100f0565b5061011792915061011b565b5090565b5b80821115610117576000815560010161011c565b600181811c9082168061014457607f821691505b6020821081141561016557634e487b7160e01b600052602260045260246000fd5b50919050565b6106558061017a6000396000f3fe608060405234801561001057600080fd5b50600436106100935760003560e01c8063313ce56711610066578063313ce5671461010357806370a082311461012257806395d89b4114610142578063a9059cbb1461014a578063dd62ed3e1461015d57600080fd5b806306fdde0314610098578063095ea7b3146100b657806318160ddd146100d957806323b872dd146100f0575b600080fd5b6100a0610188565b6040516100ad919061054c565b60405180910390f35b6100c96100c4366004610523565b610216565b60405190151581526020016100ad565b6100e260025481565b6040519081526020016100ad565b6100c96100fe3660046104e8565b610282565b6003546101109060ff1681565b60405160ff90911681526020016100ad565b6100e2610130366004610495565b60046020526000908152604090205481565b6100a06103c2565b6100c9610158366004610523565b6103cf565b6100e261016b3660046104b6565b600560209081526000928352604080842090915290825290205481565b60008054610195906105ce565b80601f01602080910402602001604051908101604052809291908181526020018280546101c1906105ce565b801561020e5780601f106101e35761010080835404028352916020019161020e565b820191906000526020600020905b8154815290600101906020018083116101f157829003601f168201915b505050505081565b3360008181526005602090815260408083206001600160a01b038716808552925280832085905551919290917f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925906102719086815260200190565b60405180910390a350600192915050565b6001600160a01b0383166000908152600460205260408120548211156102a757600080fd5b6001600160a01b03841660009081526005602090815260408083203384529091529020548211156102d757600080fd5b6001600160a01b038416600090815260046020526040812080548492906102ff9084906105b7565b90915550506001600160a01b0383166000908152600460205260408120805484929061032c90849061059f565b90915550506001600160a01b0384166000908152600560209081526040808320338452909152812080548492906103649084906105b7565b92505081905550826001600160a01b0316846001600160a01b03167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef846040516103b091815260200190565b60405180910390a35060019392505050565b60018054610195906105ce565b336000908152600460205260408120548211156103eb57600080fd5b336000908152600460205260408120805484929061040a9084906105b7565b90915550506001600160a01b0383166000908152600460205260408120805484929061043790849061059f565b90915550506040518281526001600160a01b0384169033907fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef90602001610271565b80356001600160a01b038116811461049057600080fd5b919050565b6000602082840312156104a6578081fd5b6104af82610479565b9392505050565b600080604083850312156104c8578081fd5b6104d183610479565b91506104df60208401610479565b90509250929050565b6000806000606084860312156104fc578081fd5b61050584610479565b925061051360208501610479565b9150604084013590509250925092565b60008060408385031215610535578182fd5b61053e83610479565b946020939093013593505050565b6000602080835283518082850152825b818110156105785785810183015185820160400152820161055c565b818111156105895783604083870101525b50601f01601f1916929092016040019392505050565b600082198211156105b2576105b2610609565b500190565b6000828210156105c9576105c9610609565b500390565b600181811c908216806105e257607f821691505b6020821081141561060357634e487b7160e01b600052602260045260246000fd5b50919050565b634e487b7160e01b600052601160045260246000fdfea26469706673582212201e09a3e53c90fd2b7204a5ef34d3b09d3d1fd23b1fcd35fe2b1d792c576d26dd64736f6c63430008040033"
+                "0x60806040523480156200001157600080fd5b506040516040806200203e83398101806040528101908080519060200190929190805190602001909291905050506001600f60006101000a81548160ff0219169083151502179055506040805190810160405280600681526020017f487964726f42000000000000000000000000000000000000000000000000000081525060029080519060200190620000a792919062000619565b506040805190810160405280600d81526020017f20487964726f67656e426c75650000000000000000000000000000000000000081525060039080519060200190620000f592919062000619565b506012600460006101000a81548160ff021916908360ff160217905550816000806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555080600b60006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555063a0eebb00600981905550620001b762000442640100000000026401000000009004565b600a81905550686c6b935b8bbd4000006010600073ea40d7bef6ae216c4218e9ba28f92af06cc7788673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000208190555073ea40d7bef6ae216c4218e9ba28f92af06cc7788673ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef686c6b935b8bbd4000006040518082815260200191505060405180910390a3620002cf686c6b935b8bbd400000600a54620004636401000000000262001784179091906401000000009004565b600a81905550686c6b935b8bbd400000601060007330d344806e8c13a592f54a123f560ad1976f5ec273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055507330d344806e8c13a592f54a123f560ad1976f5ec273ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef686c6b935b8bbd4000006040518082815260200191505060405180910390a3620003e7686c6b935b8bbd400000600a54620004636401000000000262001784179091906401000000009004565b600a819055506200040662000480640100000000026401000000009004565b6200041f6200059c640100000000026401000000009004565b6000600f60016101000a81548160ff0219169083151502179055505050620006c8565b6000600460009054906101000a900460ff1660ff16600a0a60095402905090565b60008282111515156200047557600080fd5b818303905092915050565b600460009054906101000a900460ff1660ff16600a0a620004df6064620004c26005600954620005bf64010000000002620017a0179091906401000000009004565b620005f364010000000002620017d1179091906401000000009004565b02600681905550600460009054906101000a900460ff1660ff16600a0a62000545606462000528605f600954620005bf64010000000002620017a0179091906401000000009004565b620005f364010000000002620017d1179091906401000000009004565b026008819055506040805190810160405280600e81526020017f302e303030303030323320455448000000000000000000000000000000000000815250600790805190602001906200059992919062000619565b50565b635bc3d880600c81905550635bcd1300600d81905550635bd64d80600e81905550565b600081830290506000831480620005e15750818382811515620005de57fe5b04145b1515620005ed57600080fd5b92915050565b600080821115156200060457600080fd5b81838115156200061057fe5b04905092915050565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f106200065c57805160ff19168380011785556200068d565b828001600101855582156200068d579182015b828111156200068c5782518255916020019190600101906200066f565b5b5090506200069c9190620006a0565b5090565b620006c591905b80821115620006c1576000816000905550600101620006a7565b5090565b90565b61196680620006d86000396000f3006080604052600436106100fc576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806306fdde0314610107578063095ea7b314610197578063141ddca7146101fc57806318160ddd1461028c57806323b872dd146102b757806327ac36c41461033c578063313ce56714610367578063521eb273146103985780636681b9fd146103ef57806370a082311461041a5780638da5cb5b1461047157806395d89b41146104c8578063a9059cbb14610558578063b0b9603b146105bd578063c8e569a8146105e8578063d4ee1d90146105ff578063dd62ed3e14610656578063ec8ac4d8146106cd575b61010533610703565b005b34801561011357600080fd5b5061011c6107f3565b6040518080602001828103825283818151815260200191508051906020019080838360005b8381101561015c578082015181840152602081019050610141565b50505050905090810190601f1680156101895780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b3480156101a357600080fd5b506101e2600480360381019080803573ffffffffffffffffffffffffffffffffffffffff16906020019092919080359060200190929190505050610891565b604051808215151515815260200191505060405180910390f35b34801561020857600080fd5b50610211610983565b6040518080602001828103825283818151815260200191508051906020019080838360005b83811015610251578082015181840152602081019050610236565b50505050905090810190601f16801561027e5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b34801561029857600080fd5b506102a1610a21565b6040518082815260200191505060405180910390f35b3480156102c357600080fd5b50610322600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190803573ffffffffffffffffffffffffffffffffffffffff16906020019092919080359060200190929190505050610a42565b604051808215151515815260200191505060405180910390f35b34801561034857600080fd5b50610351610dc6565b6040518082815260200191505060405180910390f35b34801561037357600080fd5b5061037c610dcc565b604051808260ff1660ff16815260200191505060405180910390f35b3480156103a457600080fd5b506103ad610ddf565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b3480156103fb57600080fd5b50610404610e05565b6040518082815260200191505060405180910390f35b34801561042657600080fd5b5061045b600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050610e0b565b6040518082815260200191505060405180910390f35b34801561047d57600080fd5b50610486610e54565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b3480156104d457600080fd5b506104dd610e79565b6040518080602001828103825283818151815260200191508051906020019080838360005b8381101561051d578082015181840152602081019050610502565b50505050905090810190601f16801561054a5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b34801561056457600080fd5b506105a3600480360381019080803573ffffffffffffffffffffffffffffffffffffffff16906020019092919080359060200190929190505050610f17565b604051808215151515815260200191505060405180910390f35b3480156105c957600080fd5b506105d26111b4565b6040518082815260200191505060405180910390f35b3480156105f457600080fd5b506105fd6111ba565b005b34801561060b57600080fd5b506106146113dd565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561066257600080fd5b506106b7600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050611403565b6040518082815260200191505060405180910390f35b610701600480360381019080803573ffffffffffffffffffffffffffffffffffffffff169060200190929190505050610703565b005b600080600c5442101580156107245750600f60009054906101000a900460ff165b151561072f57600080fd5b34915061073c838361148a565b610745826114da565b905061075181836114fe565b9050610768826005546116ee90919063ffffffff16565b600581905550610778838261170a565b8273ffffffffffffffffffffffffffffffffffffffff163073ffffffffffffffffffffffffffffffffffffffff167f623b3804fa71d67900d064613da8f94b9617215ee90799290593e1745087ad188484604051808381526020018281526020019250505060405180910390a36107ee34611718565b505050565b60038054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156108895780601f1061085e57610100808354040283529160200191610889565b820191906000526020600020905b81548152906001019060200180831161086c57829003601f168201915b505050505081565b600081601160003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508273ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925846040518082815260200191505060405180910390a36001905092915050565b60078054600181600116156101000203166002900480601f016020809104026020016040519081016040528092919081815260200182805460018160011615610100020316600290048015610a195780601f106109ee57610100808354040283529160200191610a19565b820191906000526020600020905b8154815290600101906020018083116109fc57829003601f168201915b505050505081565b6000600460009054906101000a900460ff1660ff16600a0a60095402905090565b6000601160008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020548211151515610acf57600080fd5b81601060008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205410151515610b1d57600080fd5b610b6f82601060008773ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205461178490919063ffffffff16565b601060008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002081905550610c0482601060008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020546116ee90919063ffffffff16565b601060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002081905550610cd682601160008773ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205461178490919063ffffffff16565b601160008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508273ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef846040518082815260200191505060405180910390a3600190509392505050565b60065481565b600460009054906101000a900460ff1681565b600b60009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b60055481565b6000601060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050919050565b6000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b60028054600181600116156101000203166002900480601f016020809104026020016040519081016040528092919081815260200182805460018160011615610100020316600290048015610f0f5780601f10610ee457610100808354040283529160200191610f0f565b820191906000526020600020905b815481529060010190602001808311610ef257829003601f168201915b505050505081565b6000808373ffffffffffffffffffffffffffffffffffffffff1614151515610f3e57600080fd5b81601060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205410151515610f8c57600080fd5b601060008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205482601060008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054011015151561101b57600080fd5b61106d82601060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205461178490919063ffffffff16565b601060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000208190555061110282601060008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020546116ee90919063ffffffff16565b601060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508273ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef846040518082815260200191505060405180910390a36001905092915050565b60085481565b6000806000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614151561121857600080fd5b6000600f60006101000a81548160ff0219169083151502179055506000600a541415156113d9576112676064611259600a80546117a090919063ffffffff16565b6117d190919063ffffffff16565b915061129260646112846014600a546117a090919063ffffffff16565b6117d190919063ffffffff16565b90506112bb6112aa82846116ee90919063ffffffff16565b600a5461178490919063ffffffff16565b600a819055506000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef846040518082815260200191505060405180910390a3600b60009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef836040518082815260200191505060405180910390a36113d86117f5565b5b5050565b600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b6000601160008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905092915050565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16141515156114c657600080fd5b600081141515156114d657600080fd5b5050565b600080624257b290506114f681846117a090919063ffffffff16565b915050919050565b6000600c5442101580156115145750600d544211155b156115ab57678ac7230489e800008210151561156a5761156361155460646115466050876117a090919063ffffffff16565b6117d190919063ffffffff16565b846116ee90919063ffffffff16565b92506115a6565b6115a36115946064611586603c876117a090919063ffffffff16565b6117d190919063ffffffff16565b846116ee90919063ffffffff16565b92505b6116e5565b600d5442101580156115bf5750600e544211155b1561165657678ac7230489e80000821015156116155761160e6115ff60646115f1603c876117a090919063ffffffff16565b6117d190919063ffffffff16565b846116ee90919063ffffffff16565b9250611651565b61164e61163f6064611631601e876117a090919063ffffffff16565b6117d190919063ffffffff16565b846116ee90919063ffffffff16565b92505b6116e4565b678ac7230489e80000821015156116a7576116a06116916064611683601e876117a090919063ffffffff16565b6117d190919063ffffffff16565b846116ee90919063ffffffff16565b92506116e3565b6116e06116d160646116c3600a876117a090919063ffffffff16565b6117d190919063ffffffff16565b846116ee90919063ffffffff16565b92505b5b5b82905092915050565b6000818301905082811015151561170457600080fd5b92915050565b611714828261182e565b5050565b600b60009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f19350505050158015611780573d6000803e3d6000fd5b5050565b600082821115151561179557600080fd5b818303905092915050565b6000818302905060008314806117c057508183828115156117bd57fe5b04145b15156117cb57600080fd5b92915050565b600080821115156117e157600080fd5b81838115156117ec57fe5b04905092915050565b611826611815670de0b6b3a7640000600a546117d190919063ffffffff16565b60095461178490919063ffffffff16565b600981905550565b6000600a5414158015611843575080600a5410155b1561190e5780601060008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508173ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef836040518082815260200191505060405180910390a361190781600a5461178490919063ffffffff16565b600a819055505b6000600a54111515611936576001600f60016101000a81548160ff0219169083151502179055505b50505600a165627a7a72305820db2b321557a266a5539226ca32bfce96ebe614cf91c7bea11b31e19c6ba2ff24002900000000000000000000000063cf35c9111f65ba07ec7355a744bd0113fcf668000000000000000000000000164ce394ba9e968f1c3a88a1ac68d62e09781ac2"
             )
         elif address == BENIGN_CONTRACT:
             return HexBytes(
